+++ packages/components/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `CopyToClipboard` component
 - `TreeArtificial`
 
 ### Changed
@@ -35,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `FadeIn` component
+- `FadeIn`
 
 ### Changed